// src/app/modules/home/interfaces/user-model.interface.ts
export interface UserModel {
  price?: number;
  gymName?: string;
  id?: string;
  email?: string;
  initialTime?: number | any;
  idUser?: string;
  maximumNumberOfPeople?: number;
  numberOfPeople?: number;
  paymentAmount?: number;
  pricePerHour?: number;
  reservationDate?: string | any;
  endTime?: number | any;
}